Moved to ResultActivity
The brightnessrecognition is now called inside ResultActivity and the thread is no longer anonimous. Added strings  to the resources to support localization of the toasts' text
